worker/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pytz==2024.1
PyMuPDF>=1.23.0
Pillow>=10.0.0
google-genai
python-dotenv>=1.0.0